mao/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mao-cli"
version = "0.1.0"
description = "MAO - Multi-Agent Orchestration Testing Platform CLI and MCP Server"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "MAO Team"}
]
keywords = ["ai", "agents", "testing", "mcp", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "rich>=13.7",
    "pyyaml>=6.0",
    "backoff>=2.0",
]

[project.optional-dependencies]
mcp = ["mcp>=1.0"]
security = ["cryptography>=41.0", "keyring>=24.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
]

[project.scripts]
mao = "mao.cli.main:cli"

[project.urls]
Homepage = "https://github.com/mao-testing/mao"
Documentation = "https://docs.mao-testing.com"
Repository = "https://github.com/mao-testing/mao"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
